css/css-borders-4/parsing/box-shadow-position-computed.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Borders 4 Test: Computed values of 'box-shadow-position'</title>
  <link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
  <link rel="help" href="https://drafts.csswg.org/css-borders-4/#box-shadow-position">
  <meta name="assert" content="Computed value of 'box-shadow-position' is correct">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
  <style>
    #target {
      font-size: 20px;
    }
  </style>
</head>
<body>
  <div id="target"></div>
  <script>
    test_computed_value("box-shadow-position", "inset");
    test_computed_value("box-shadow-position", "outset");
    test_computed_value("box-shadow-position", "inset, outset");
  </script>
</body>
</html>
